move all target handling into blocknode.go
